# docs/swagger.yaml
definitions:
  dto.OpeningRequest:
    properties:
      company:
        type: string
      link:
        type: string
      location:
        type: string
      remote:
        type: boolean
      role:
        type: string
      salary:
        type: number
    type: object
  response.ErrorResponse:
    properties:
      errorCode:
        type: integer
      message:
        type: string
    type: object
  response.SuccessResponse:
    properties:
      data:
        $ref: '#/definitions/schemas.OpeningResponse'
      message:
        type: string
    type: object
  schemas.OpeningResponse:
    properties:
      company:
        type: string
      create_at:
        type: string
      delete_at:
        type: string
      id:
        type: integer
      link:
        type: string
      location:
        type: string
      remote:
        type: boolean
      role:
        type: string
      salary:
        type: number
      update_at:
        type: string
    type: object
info:
  contact: {}
paths:
  /opening:
    post:
      consumes:
      - application/json
      description: Create a new opening
      parameters:
      - description: Request body
        in: body
        name: opening
        required: true
        schema:
          $ref: '#/definitions/dto.OpeningRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Create a new opening
      tags:
      - openings
  /opening/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a opening
      parameters:
      - description: Opening ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Delete a opening
      tags:
      - openings
    get:
      consumes:
      - application/json
      description: get a opening
      parameters:
      - description: Opening ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: get a opening
      tags:
      - openings
    put:
      consumes:
      - application/json
      description: Update a opening
      parameters:
      - description: Opening ID
        in: path
        name: id
        required: true
        type: string
      - description: Request body
        in: body
        name: opening
        required: true
        schema:
          $ref: '#/definitions/dto.OpeningRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: Update a opening
      tags:
      - openings
  /openings:
    get:
      consumes:
      - application/json
      description: List all openings
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/response.SuccessResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/response.ErrorResponse'
      summary: List all openings
      tags:
      - openings
swagger: "2.0"